ENT-12650 Use std::filesystem by default (#5)
To help move forward switch to using C++17 as target and use std
filesystem that comes with it. Add USE_BOOST_FILESYSTEM that will use
boost::filesystem (as previous) in case, but that defaults OFF. Note
change in API to pass paths by std::string so whichever filesystem is in
use does not bother the calling code.